Fix a race and a leak in the cache consumer.
The race is that if we change out the consume_callback while
the previous thread is still running, we could end up in a state
where the callback was half-updated (as far as I can tell,
swapping a std::function is *not* atomic).

The leak is that if the previous thread had not been explicitly
close() before change_consume_callback() is called, it would
leak the previous thread and forget about it.

But it turns out that we don't need to kill the thread at all
unless we are doing a close.  Lock a mutex around the
consume_callback_ member variable, and we can ensure that it
won't change while the thread is using it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>